+++ src/modules/rooms/dto/responses/index.ts
@@ -1,2 +1,4 @@
 export * from './created-room.dto';
 export * from './room-details.dto';
+export * from './room-list-item.dto';
+export * from './paginated-rooms.dto';