+++ src/modules/rooms/swagger-docs/index.ts
@@ -1,2 +1,3 @@
 export * from './create-room';
+export * from './get-rooms';
 export * from './get-room';